Fixed a few compile warnings

+++ libzhl/LuaCore.cpp
@@ -373,7 +373,7 @@ namespace lua {
 		int ToInteger(lua_State* L, int stackPosition) {
 			lua_pushinteger(L, stackPosition);
 			lua_gettable(L, -2);
-			int res = luaL_checkinteger(L, -1);
+			int res = (int)luaL_checkinteger(L, -1);
 			lua_pop(L, 1);
 			return res;
 		}
